requirements.txt
docx2python==3.6.2
langchain==1.2.13
faiss-cpu==1.13.2
transformers==5.3.0
sentence-transformers==5.3.0
langchain-ollama==1.0.1
langchain-huggingface==1.2.1
langchain-community==0.4.1
langgraph==1.1.3
bs4==0.0.2
pypdf==6.9.1